Refactor batch jobs to not need Batchable and QueryBatch
