+++ yii2-adapter/legacy/base/Batchable.php
@@ -15,6 +15,7 @@ use Countable;
  *
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @since 4.4.0
+ * @deprecated 6.0.0
  */
 interface Batchable extends Countable
 {
