+++ yii2-adapter/legacy/db/QueryBatcher.php
@@ -21,6 +21,7 @@ use yii\db\QueryInterface;
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
  *
  * @since 4.4.0
+ * @deprecated 6.0.0
  */
 class QueryBatcher implements Batchable
 {
